feat: 유저 섹션 분리
