--- client/presentation/src/main/java/com/andone/memorip/presentation/screen/user/component/ProfileImgageButton.kt
@@ -1,2 +1,0 @@
-package com.andone.memorip.presentation.screen.user.component
-
